added flower celery and redis
